LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/xmlhttprequest-sync-block-scripts-expected.txt

CONSOLE MESSAGE: line 17: XMLHttpRequest cannot load data:,. Cross origin requests are only supported for HTTP.

FAIL Check that while a sync XHR is in flight async script loads don't complete and run script NetworkError: DOM Exception 19
